Ressurecting my laptop for linux dev
Previously I fought Arch linux all the way down. Considering the RasPi
is a Debian based distro, going back to Ubuntu so it just works is
probably fine. That said I'm taking the LXQt flavor of it so it reports
as Debian in Ansible. Meh. Minimal additions allowed me to get to the
same dev env as the RasPi in under an hour. I'll call that a win.